requirements.txt
mod-wsgi ~= 4.5.24
Django ~= 2.0
django-aliyun-oss2-storage ~= 0.1.5
django-redis-cache ~= 1.7.1
djangorestframework ~= 3.7.3
oss2 ~= 2.4.0
psycopg2 ~= 2.7.3.2
Scrapy ~= 1.5.0
scrapy-djangoitem ~= 1.1.1
django-webpack-loader ~= 0.5.0
django-allauth ~= 0.34.0
django-rest-auth ~= 0.9.2
django-crispy-forms ~= 1.7.0
celery[redis] ~= 4.1.0
django-celery-beat ~= 1.1.0
flower ~= 0.9.2
django-baton ~= 1.0.8
